requirements.txt
bs4
requests
python-vlc
RPi.gpio

